Trait postgres_binary_copy::StreamingIterator

pub trait StreamingIterator {
    type Item: ?Sized;
    fn next(&mut self) -> Option<&Self::Item>;
}

Like Iterator, except that it returns borrowed values.

In contrast to Iterator<Item = &T>, a type implementing StreamingIterator<Item = T> does not need to have all of the values it returns in memory at the same time.

All Iterators over &T are also StreamingIterators over T.

Associated Types

type Item: ?Sized

The type of elements being iterated.

Required Methods

fn next(&mut self) -> Option<&Self::Item>

Advances the iterator and returns the next value.

Returns None when the end is reached.